Separate Peer Server logic
