+++ src/main/kotlin/brs/api/http/ApiServer.kt
@@ -1,0 +1,5 @@
+package brs.api.http
+
+interface ApiServer {
+    fun shutdown()
+}
