+++ src/main/kotlin/brs/peer/PeerServer.kt
@@ -1,0 +1,5 @@
+package brs.peer
+
+interface PeerServer {
+    fun shutdown()
+}
